requirements.txt
azure-common==1.1.28
azure-core==1.33.0
azure-identity==1.21.0
azure-mgmt-core==1.5.0
azure-mgmt-resourcegraph==8.0.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
idna==3.10
isodate==0.7.2
markdown-it-py==3.0.0
mdurl==0.1.2
msal==1.32.0
msal-extensions==1.3.1
msrest==0.7.1
numpy==2.2.4
oauthlib==3.2.2
pandas==2.2.3
pycparser==2.22
pyfiglet==1.0.2
Pygments==2.19.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
shellingham==1.5.4
six==1.17.0
tabulate==0.9.0
termcolor==2.3.0
typer==0.15.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
yaspin==3.1.0
